requirements.txt
Flask==2.2.3
SQLAlchemy==2.0.4
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==0.15.1
Werkzeug==2.2.3
Pillow==8.3.1
python-dotenv==0.21.1
Flask-Migrate==3.1.0
pymysql==1.0.2
email-validator==1.3.1
psycopg2-binary==2.9.5
Faker==18.9.0
Jinja2==3.1.2
pytest==7.3.1
pytest-flask==1.2.0
python-dateutil==2.8.2
argparse>=1.4.0
WTForms==3.2.1
dnspython==2.7.0